Update ref graphs; Move tests to common; Fix pylint
