+++ nncf/torch/graph/patterns.py
@@ -94,7 +94,8 @@ class GraphPatternFactory:
 
         ANY_BN_ACT_COMBO = BN + ACTIVATIONS | ACTIVATIONS + BN | BN | ACTIVATIONS
 
-        FULL_PATTERN_GRAPH = LINEAR_OPS + ANY_BN_ACT_COMBO | ANY_BN_ACT_COMBO | ARITHMETIC + ANY_BN_ACT_COMBO
+        FULL_PATTERN_GRAPH = LINEAR_OPS + ANY_BN_ACT_COMBO | BN + ACTIVATIONS | ACTIVATIONS + BN | \
+                             ARITHMETIC + ANY_BN_ACT_COMBO
 
         return FULL_PATTERN_GRAPH
